(fix) 2. Парковка машин [#853 #329644]

+++ src/main/java/ru/job4j/design/lsp/Auto.java
@@ -1,5 +1,5 @@
 package ru.job4j.design.lsp;
 
 public interface Auto {
-    float getSize();
+    int getSize();
 }
